Move Testbench-based tests to tests/Integration
